:recycle:(address-manager): merge token-refresh-job and ttl-refresher-job into parameterized RefreshJob<T>
